+++ src/source/kisskh.ts
@@ -1,6 +1,7 @@
 import axios from "axios";
 import * as cheerio from "cheerio";
 import {
+  Args,
   ContentType,
   MetaDetail,
   MetaPreview,
@@ -8,7 +9,7 @@ import {
   Stream,
   Subtitle,
 } from "stremio-addon-sdk";
-import { Prefix } from "../lib/manifest.js";
+import { Prefix, UserConfig } from "../lib/manifest.js";
 import { axiosGet } from "../utils/axios.js";
 import { cache } from "../utils/cache.js";
 import { envGet } from "../utils/env.js";
@@ -97,12 +98,16 @@ class KissKHScraperr extends BaseProvider {
     tv: "1",
   };
   private tokenJsCode: string | null = null;
+  private nsfwIds = new Set([
+    12519, 12518, 12517, 12516, 12515, 12514, 12513, 12510, 12504, 12503, 12495,
+    12491, 12480, 12413, 12378, 12332, 12331, 12330, 12314, 12285, 12284, 12200,
+    12179, 12177, 12127, 12125, 12124, 12123, 12106, 11915, 11834, 11782, 11519,
+    11518, 11517, 11511, 11509, 11436, 10942, 10761,
+  ]);
 
-  async searchCatalog(
-    id: string,
-    type: ContentType,
-    search: string,
-  ): Promise<MetaPreview[]> {
+  async searchCatalog(args: Args, config: UserConfig): Promise<MetaPreview[]> {
+    const { id, type, extra } = args;
+    const search = extra.search;
     this.logger.log(`Search | ${search}`);
     const searchResults = await this.searchContent(
       search,
@@ -115,6 +120,11 @@ class KissKHScraperr extends BaseProvider {
     const filterResults = searchResults.filter((result) => {
       if (type == "series") return result.episodesCount > 1;
       else return result.episodesCount == 1;
+    });
+    filterResults.forEach((result) => {
+      if (!config.nsfw && this.nsfwIds.has(parseInt(result.id))) {
+        result.thumbnail = this.nsfwDefaultThumbnail;
+      }
     });
     const metas = filterResults.map((detail) => {
       const meta: MetaPreview = {
@@ -137,11 +147,9 @@ class KissKHScraperr extends BaseProvider {
    * @param skip
    * @returns
    */
-  async getCatalog(
-    id: string,
-    type: ContentType,
-    skip?: number,
-  ): Promise<MetaPreview[]> {
+  async getCatalog(args: Args, config: UserConfig): Promise<MetaPreview[]> {
+    const { id, type, extra } = args;
+    const skip = extra.skip;
     let t = this.TYPE[type];
     const ongoing = 1;
     const completed = 2;
@@ -171,19 +179,23 @@ class KissKHScraperr extends BaseProvider {
     );
     const promises = urls.map(async (url) => {
       this.logger.log(`GET catalog | ${url}`);
-      return axiosGet(url);
+      return axiosGet<KisskhCatalogData>(url);
     });
-    const datas: KisskhCatalogData[] = await Promise.all(promises);
+    let datas = await Promise.all(promises);
     const metas = datas
       .map((data) => {
-        // const data: KisskhCatalogData = data.data;
+        if (!data) return [];
         const metas = data.data.map((kisskhMeta) => {
+          let thumbnail = kisskhMeta.thumbnail;
+          if (!config.nsfw && this.nsfwIds.has(kisskhMeta.id)) {
+            thumbnail = this.nsfwDefaultThumbnail;
+          }
           const meta: MetaPreview = {
             id: `${Prefix.KISSKH}:${kisskhMeta.id}`,
             name: kisskhMeta.title,
             type: type,
             background: kisskhMeta.thumbnail,
-            poster: kisskhMeta.thumbnail,
+            poster: thumbnail,
             posterShape: "regular",
           };
           return meta;
@@ -242,14 +254,10 @@ class KissKHScraperr extends BaseProvider {
   }
 
   async getStreams(
-    title: string,
-    type: ContentType,
-    year?: number,
-    season?: number,
-    episode?: number,
-    id?: string,
-    altTitle?: string,
-  ): Promise<Stream[] | null> {
+    content: ContentDetail,
+    config: UserConfig,
+  ): Promise<Stream[]> {
+    const { title, type, year, season, episode, id, altTitle } = content;
     try {
       if (id) {
         const streamKey = `streams:kisskh:${type}:${id}:${season}:${episode}`;
@@ -272,22 +280,19 @@ class KissKHScraperr extends BaseProvider {
       );
       if (!searchResult[0]) {
         this.logger.log("No results");
-        return null;
+        return [];
       }
       const streams = await this.generateStreamsAndSubtitles(
         searchResult[0],
-        type,
-        year,
-        season,
-        episode,
-        id,
+        content,
+        config,
       );
-      cache.set(streamKey, streams, 4 * 60 * 60 * 1000);
+      if (streams) cache.set(streamKey, streams, 4 * 60 * 60 * 1000);
       return streams;
     } catch (error: any) {
       this.logger.error(`Error | ${error.message}`);
+      return [];
     }
-    return null;
   }
 
   async getSubtitles(content: ContentDetail): Promise<Subtitle[]> {
@@ -335,42 +340,36 @@ class KissKHScraperr extends BaseProvider {
 
   async generateStreamsAndSubtitles(
     searchResult: SearchResult,
-    type: ContentType,
-    year?: number,
-    season?: number,
-    episode?: number,
-    id?: string,
-  ) {
+    content: ContentDetail,
+    config: UserConfig,
+  ): Promise<Stream[]> {
+    const { episode, id, season, year, type } = content;
     const episodeId = await this._getEpisode(searchResult.id, type, episode);
     const token = await this._getToken(episodeId, this.viGuid);
     const stream = await this._getStream(episodeId, token);
-    const url = this._fixUrl(stream.Video!);
-    let info;
-    try {
-      info = parseStreamInfo(url);
-    } catch (error) {
-      this.logger.error(`Fail to parse stream info | ${error}`);
-    }
+    if (!stream) return [];
+    if (!stream.Video) return [];
+    const url = this._fixUrl(stream.Video);
     const subtitleKey = `subtitles:kisskh:${type}:${id}:${season}:${episode}`;
     // Handle subtitles
     let subtitles = cache.get(subtitleKey);
     if (subtitles) cache.set(subtitleKey, subtitles);
     else {
       subtitles = await this._getSubtitles(episodeId);
-      cache.set(subtitleKey, subtitles);
+      if (subtitles) cache.set(subtitleKey, subtitles);
     }
-
+    const info = config.info ? await parseStreamInfo(url) : undefined;
     const formatTitle = this.formatStreamTitle(
       searchResult.title,
       year,
       season,
       episode,
-      await info,
+      info,
     );
     const streams: Stream[] = [
       {
         url: url,
-        name: "yastream",
+        name: this.displayName,
         title: formatTitle,
         behaviorHints: {
           notWebReady: true,
@@ -436,8 +435,11 @@ class KissKHScraperr extends BaseProvider {
   public async getDetail(kisskhId: string): Promise<KisskhDetail> {
     const url = `${this.detailUrl}${kisskhId}`;
     this.logger.log(`GET detail | ${url}`);
-    const episodesData = await axiosGet(url, { headers: this.headers });
-    return episodesData;
+    const episodesData = await axiosGet<KisskhDetail>(url, {
+      headers: this.headers,
+    });
+    if (episodesData) return episodesData;
+    else throw new Error(`Not found detail from id | ${kisskhId}`);
   }
 
   private async _getEpisode(
@@ -477,7 +479,8 @@ class KissKHScraperr extends BaseProvider {
   private async _getStream(episodeId: string, token: string) {
     const url = this.episodeUrl.replace("{id}", episodeId) + token;
     this.logger.log(`GET stream | ${url}`);
-    const stream = await axiosGet(url);
+    const stream = await axiosGet<StreamResponse>(url);
+    if (!stream) return null;
     this.logger.log(`Stream Url | ${stream.Video}`);
     return stream;
   }
@@ -486,7 +489,8 @@ class KissKHScraperr extends BaseProvider {
     const token = await this._getToken(episodeId, this.subGuid);
     const subtitleUrl = this.subUrl.replace("{id}", episodeId) + token;
     this.logger.log(`GET subtitles | ${subtitleUrl}`);
-    const subtitleDatas: SubResponse[] = await axiosGet(subtitleUrl);
+    const subtitleDatas = await axiosGet<SubResponse[]>(subtitleUrl);
+    if (!subtitleDatas) return [];
     const subtitles: Subtitle[] = [];
     for (const [index, subtitleData] of subtitleDatas.entries()) {
       const lang = iso639FromCountryCode(subtitleData.land as CountryCode);